+++ database/migrations/2020_09_08_134956_create_table_total_pontos.php
@@ -14,8 +14,8 @@ class CreateTableTotalPontos extends Migration
     public function up()
     {
         Schema::create('total_pontos', function (Blueprint $table) {
-            $table->integer('idTemporada');
-            $table->integer('idTime');
+            $table->unsignedBigInteger('idTemporada');
+            $table->unsignedBigInteger('idTime');
             $table->integer('numPartida');
             $table->integer('pontos');
             $table->integer('vitorias');
